requirements.txt
certifi==2021.5.30
ftfy==6.0.3
numpy==1.21.2
Pillow==8.3.1
regex==2021.8.28
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
wcwidth==0.2.5
